Fixed objc push notification fail

+++ Shared/BrainCloudPushNotification.mm
@@ -159,7 +159,7 @@
                             fcmContent:(NSString *)fcmContent
                             iosContent:(NSString *)iosContent
                        facebookContent:(NSString *)facebookContent
-                             startTimeUTC:(double)startTimeUTC
+                             startTimeUTC:(int64_t)startTimeUTC
                        completionBlock:(BCCompletionBlock)cb
                   errorCompletionBlock:(BCErrorCompletionBlock)ecb
                               cbObject:(BCCallbackObject)cbObject
@@ -282,7 +282,7 @@
 - (void)scheduleNormalizedPushNotificationUTC:(NSString *)toProfileId
                              alertContentJson:(NSString *)alertContentJson
                                customDataJson:(NSString *)customDataJson
-                                    startTimeUTC:(double)startTimeUTC
+                                    startTimeUTC:(int64_t)startTimeUTC
                               completionBlock:(BCCompletionBlock)cb
                          errorCompletionBlock:(BCErrorCompletionBlock)ecb
                                      cbObject:(BCCallbackObject)cbObject
@@ -317,7 +317,7 @@
 - (void)scheduleRichPushNotificationUTC:(NSString *)toProfileId
                  notificationTemplateId:(int)notificationTemplateId
                        substitutionJson:(NSString *)substitutionJson
-                              startTimeUTC:(double)startTimeUTC
+                              startTimeUTC:(int64_t)startTimeUTC
                         completionBlock:(BCCompletionBlock)cb
                    errorCompletionBlock:(BCErrorCompletionBlock)ecb
                                cbObject:(BCCallbackObject)cbObject
